Move policies to own folder, fix C&P bugs
